refactor: rename processor registering classes
These classes had Event appended to them. As these processors process commands this is misleading and could cause
confusion.
