+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/processing/user/UserProcessors.java
@@ -5,34 +5,30 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.engine.processing.identity;
+package io.camunda.zeebe.engine.processing.user;
 
 import io.camunda.zeebe.engine.processing.distribution.CommandDistributionBehavior;
 import io.camunda.zeebe.engine.processing.streamprocessor.TypedRecordProcessors;
 import io.camunda.zeebe.engine.processing.streamprocessor.writers.Writers;
 import io.camunda.zeebe.engine.state.mutable.MutableProcessingState;
 import io.camunda.zeebe.protocol.record.ValueType;
-import io.camunda.zeebe.protocol.record.intent.AuthorizationIntent;
+import io.camunda.zeebe.protocol.record.intent.UserIntent;
 import io.camunda.zeebe.stream.api.state.KeyGenerator;
 
-public final class AuthorizationEventProcessors {
-
-  public static void addAuthorizationProcessors(
+public class UserProcessors {
+  public static void addUserProcessors(
       final KeyGenerator keyGenerator,
       final TypedRecordProcessors typedRecordProcessors,
       final MutableProcessingState processingState,
       final Writers writers,
       final CommandDistributionBehavior distributionBehavior) {
     typedRecordProcessors.onCommand(
-        ValueType.AUTHORIZATION,
-        AuthorizationIntent.CREATE,
-        new AuthorizationCreateProcessor(
-            keyGenerator, processingState, writers, distributionBehavior));
-
+        ValueType.USER,
+        UserIntent.CREATE,
+        new UserCreateProcessor(keyGenerator, processingState, writers, distributionBehavior));
     typedRecordProcessors.onCommand(
-        ValueType.AUTHORIZATION,
-        AuthorizationIntent.ADD_PERMISSION,
-        new AuthorizationAddPermissionProcessor(
-            writers, keyGenerator, processingState, distributionBehavior));
+        ValueType.USER,
+        UserIntent.UPDATE,
+        new UserUpdateProcessor(keyGenerator, processingState, writers, distributionBehavior));
   }
 }
